+++ src/app/modules/portfolio/components/tradeable-asset-sum-card/tradeable-asset-sum-card.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input, ChangeDetectionStrategy } from '@angular/core';
+import { Component, OnInit, Input, ChangeDetectionStrategy, OnChanges, SimpleChanges } from '@angular/core';
 
 import { AssetOverview } from '../../models/asset-overview';
 
@@ -11,15 +11,32 @@ import { AssetOverview } from '../../models/asset-overview';
   styleUrls: ['./tradeable-asset-sum-card.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class TradeableAssetSumCardComponent implements OnInit {
+export class TradeableAssetSumCardComponent implements OnInit, OnChanges {
 
   @Input() baseCurrency: string;
   @Input() overview: AssetOverview;
   @Input() title: string;
+  grossPortfolioValue: number;
+  currentDebtRatio: number;
 
   constructor() { }
 
   ngOnInit() {
+    this.updateValues();
+  }
+
+  ngOnChanges(changes: SimpleChanges) {
+    if (changes.overview) {
+      this.updateValues();
+    }
+  }
+
+  updateValues() {
+    this.grossPortfolioValue = this.overview.currentValue - this.overview.debtValue;
+    if (this.grossPortfolioValue) {
+      this.currentDebtRatio = Math.abs(this.overview.debtValue) / this.grossPortfolioValue;
+    }
+
   }
 
 }